requirements.txt
Flask
Flask-Cors
python-docx